IPv6 support on SSL Validation handler
Modules run on a ipv6-only machine will fail with SSL error when connecting to https urls:
"Failed to validate the SSL certificate for %s:%s. Make sure your managed systems have a valid CA certificate installed.  If the website serving the url uses SNI you need python >= 2.7.9 on your managed machine.  You can use validate_certs=False if you do not need to confirm the server\s identity but this is unsafe and not recommended"
(which is totally misleading, but thats another story). 
It turns out that the socket.connect method only works for either ipv4 or ipv6. socket.create_connection on the other hand works for both.